+++ backend/src/main/java/com/team4/backend/repository/FileMetaDataRepository.java
@@ -1,16 +1,11 @@
 package com.team4.backend.repository;
 
-import com.team4.backend.model.FileMetaData;
-import org.springframework.data.domain.Pageable;
+import com.team4.backend.model.FileMetadata;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
 
 @Repository
-public interface FileMetaDataRepository extends ReactiveMongoRepository<FileMetaData, String> {
-    Mono<Long> countAllByIsValidFalseAndIsSeenFalse();
-
-    Flux<FileMetaData> findAllByIsValidFalseAndIsSeenFalse(Pageable pageable);
+public interface FileMetadataRepository extends ReactiveMongoRepository<FileMetadata, String> {
+    Flux<FileMetadata> findFileMetadataById(String id);
 }
