+++ frontend/src/components/ListCvInternshipManagerView.js
@@ -13,7 +13,7 @@ const ListCvInternshipManagerView = () => {
     const getNbrCvs = () => {
       axios({
         method: "GET",
-        url: "http://localhost:8080/fileMetaData/countAllInvalidCvNotSeen/",
+        url: "http://localhost:8080/file/countAllInvalidCvNotSeen/",
         headers: {
           Authorization: sessionStorage.getItem("jwt"),
         },
@@ -30,7 +30,7 @@ const ListCvInternshipManagerView = () => {
     const getCvs = (newPage) => {
       axios({
         method: "GET",
-        url: `http://localhost:8080/fileMetaData/getListInvalidCvNotSeen/${newPage}`,
+        url: `http://localhost:8080/file/getListInvalidCvNotSeen/${newPage}`,
         headers: {
           Authorization: sessionStorage.getItem("jwt"),
         },
